requirements.txt
streamlit==1.50.0
opencv-python
pillow
torch
torchvision
matplotlib
numpy
pytesseract
requests